Refactor parser Services
