+++ src/main/java/com/controller/io/parser/CommandParserService.java
@@ -1,4 +1,4 @@
-package com.controller.io;
+package com.controller.io.parser;
 
 import com.controller.io.validator.PlaceCommandValidatorService;
 import com.controller.io.validator.StringInputValidatorService;
