+++ src/main/java/com/controller/io/parser/InputArgParserService.java
@@ -1,4 +1,4 @@
-package com.controller.io;
+package com.controller.io.parser;
 
 import org.apache.commons.cli.*;
